Move Nillable, Reference and Referenceable from ereporting to api
