+++ core/api/src/main/java/org/n52/sos/w3c/Nillable.java
@@ -26,7 +26,7 @@
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
  * Public License for more details.
  */
-package org.n52.sos.util;
+package org.n52.sos.w3c;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
@@ -41,9 +41,7 @@ import com.google.common.base.Optional;
  *
  * @author Christian Autermann
  * @param <T> the instance type
- * @deprecated {@link org.n52.sos.w3c.Nillable}
  */
-@Deprecated
 public abstract class Nillable<T> {
 
     public Nillable() {
