+++ core/api/src/main/java/org/n52/sos/w3c/xlink/Referenceable.java
@@ -26,22 +26,16 @@
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
  * Public License for more details.
  */
-package org.n52.sos.util;
+package org.n52.sos.w3c.xlink;
 
 import java.net.URI;
+
+import org.n52.sos.w3c.Nillable;
 
 import com.google.common.base.Function;
 import com.google.common.base.Objects;
 import com.google.common.base.Preconditions;
 
-/**
- * @author <a href="mailto:c.hollmann@52north.org">Carsten Hollmann</a>
- * @since
- *
- * @param <T>
- * @deprecated {@link org.n52.sos.w3c.xlink.Reference}
- */
-@Deprecated
 public abstract class Referenceable<T> {
 
     public abstract Reference getReference();
